google_api_storage v0.1.0 GoogleApi.Storage.V1.Api.DefaultObjectAccessControls

API calls for all endpoints tagged DefaultObjectAccessControls.

Functions

storage_default_object_access_controls_delete(connection, bucket, entity, opts \\ [])

storage_default_object_access_controls_delete(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, nil} | {:error, Tesla.Env.t()}

Permanently deletes the default object ACL entry for the specified entity on the specified bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a bucket.

entity (String.t): The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %{}} on success {:error, info} on failure

storage_default_object_access_controls_get(connection, bucket, entity, opts \\ [])

storage_default_object_access_controls_get(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.ObjectAccessControl.t()}
  | {:error, Tesla.Env.t()}

Returns the default object ACL entry for the specified entity on the specified bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a bucket.

entity (String.t): The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.ObjectAccessControl{}} on success {:error, info} on failure

storage_default_object_access_controls_insert(connection, bucket, opts \\ [])

storage_default_object_access_controls_insert(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.ObjectAccessControl.t()}
  | {:error, Tesla.Env.t()}

Creates a new default object ACL entry on the specified bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a bucket.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (ObjectAccessControl):

Returns

{:ok, %GoogleApi.Storage.V1.Model.ObjectAccessControl{}} on success {:error, info} on failure

storage_default_object_access_controls_list(connection, bucket, opts \\ [])

storage_default_object_access_controls_list(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.ObjectAccessControls.t()}
  | {:error, Tesla.Env.t()}

Retrieves default object ACL entries on the specified bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a bucket.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:ifMetagenerationMatch (String.t): If present, only return default ACL listing if the bucket's current metageneration matches this value.

:ifMetagenerationNotMatch (String.t): If present, only return default ACL listing if the bucket's current metageneration does not match the given value.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.ObjectAccessControls{}} on success {:error, info} on failure

storage_default_object_access_controls_patch(connection, bucket, entity, opts \\ [])

storage_default_object_access_controls_patch(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.ObjectAccessControl.t()}
  | {:error, Tesla.Env.t()}

Patches a default object ACL entry on the specified bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a bucket.

entity (String.t): The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (ObjectAccessControl):

Returns

{:ok, %GoogleApi.Storage.V1.Model.ObjectAccessControl{}} on success {:error, info} on failure

storage_default_object_access_controls_update(connection, bucket, entity, opts \\ [])

storage_default_object_access_controls_update(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.ObjectAccessControl.t()}
  | {:error, Tesla.Env.t()}

Updates a default object ACL entry on the specified bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a bucket.

entity (String.t): The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (ObjectAccessControl):

Returns

{:ok, %GoogleApi.Storage.V1.Model.ObjectAccessControl{}} on success {:error, info} on failure