google_api_storage v0.1.0 GoogleApi.Storage.V1.Api.Notifications

API calls for all endpoints tagged Notifications.

Functions

storage_notifications_delete(connection, bucket, notification, opts \\ [])

storage_notifications_delete(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, nil} | {:error, Tesla.Env.t()}

Permanently deletes a notification subscription.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): The parent bucket of the notification.

notification (String.t): ID of the notification to delete.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %{}} on success {:error, info} on failure

storage_notifications_get(connection, bucket, notification, opts \\ [])

storage_notifications_get(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Notification.t()} | {:error, Tesla.Env.t()}

View a notification configuration.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): The parent bucket of the notification.

notification (String.t): Notification ID

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.Notification{}} on success {:error, info} on failure

storage_notifications_insert(connection, bucket, opts \\ [])

storage_notifications_insert(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Notification.t()} | {:error, Tesla.Env.t()}

Creates a notification subscription for a given bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): The parent bucket of the notification.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Notification):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Notification{}} on success {:error, info} on failure

storage_notifications_list(connection, bucket, opts \\ [])

storage_notifications_list(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Notifications.t()} | {:error, Tesla.Env.t()}

Retrieves a list of notification subscriptions for a given bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of a Google Cloud Storage bucket.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.Notifications{}} on success {:error, info} on failure