google_api_storage v0.1.0 GoogleApi.Storage.V1.Api.Objects

API calls for all endpoints tagged Objects.

Functions

storage_objects_compose(connection, destination_bucket, destination_object, opts \\ [])

storage_objects_compose(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Concatenates a list of existing objects into a new object in the same bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

destination_bucket (String.t): Name of the bucket in which to store the new object.

destination_object (String.t): Name of the new object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:destinationPredefinedAcl (String.t): Apply a predefined set of access controls to the destination object.

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:kmsKeyName (String.t): Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object. Overrides the object metadata's kms_key_name value, if any.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (ComposeRequest):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_copy(connection, source_bucket, source_object, destination_bucket, destination_object, opts \\ [])

storage_objects_copy(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Copies a source object to a destination object. Optionally overrides metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

source_bucket (String.t): Name of the bucket in which to find the source object.

source_object (String.t): Name of the source object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

destination_bucket (String.t): Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

destination_object (String.t): Name of the new object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:destinationPredefinedAcl (String.t): Apply a predefined set of access controls to the destination object.

:ifGenerationMatch (String.t): Makes the operation conditional on whether the destination object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the destination object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the destination object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the destination object's current metageneration does not match the given value.

:ifSourceGenerationMatch (String.t): Makes the operation conditional on whether the source object's current generation matches the given value.

:ifSourceGenerationNotMatch (String.t): Makes the operation conditional on whether the source object's current generation does not match the given value.

:ifSourceMetagenerationMatch (String.t): Makes the operation conditional on whether the source object's current metageneration matches the given value.

:ifSourceMetagenerationNotMatch (String.t): Makes the operation conditional on whether the source object's current metageneration does not match the given value.

:projection (String.t): Set of properties to return. Defaults to noAcl, unless the object resource specifies the acl property, when it defaults to full.

:sourceGeneration (String.t): If present, selects a specific revision of the source object (as opposed to the latest version, the default).

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Object):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_delete(connection, bucket, object, opts \\ [])

storage_objects_delete(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, nil} | {:error, Tesla.Env.t()}

Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, permanently deletes a specific revision of this object (as opposed to the latest version, the default).

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %{}} on success {:error, info} on failure

storage_objects_get(connection, bucket, object, opts \\ [])

storage_objects_get(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Retrieves an object or its metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, selects a specific revision of this object (as opposed to the latest version, the default).

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:projection (String.t): Set of properties to return. Defaults to noAcl.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_get_iam_policy(connection, bucket, object, opts \\ [])

storage_objects_get_iam_policy(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, GoogleApi.Storage.V1.Model.Policy.t()} | {:error, Tesla.Env.t()}

Returns an IAM policy for the specified object.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, selects a specific revision of this object (as opposed to the latest version, the default).

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.Policy{}} on success {:error, info} on failure

storage_objects_insert(connection, bucket, opts \\ [])

storage_objects_insert(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Stores a new object and metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:contentEncoding (String.t): If set, sets the contentEncoding property of the final object to this value. Setting this parameter is equivalent to setting the contentEncoding metadata property. This can be useful when uploading an object with uploadType=media to indicate the encoding of the content being uploaded.

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:kmsKeyName (String.t): Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object. Overrides the object metadata's kms_key_name value, if any. Limited availability; usable only by enabled projects.

:name (String.t): Name of the object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

:predefinedAcl (String.t): Apply a predefined set of access controls to this object.

:projection (String.t): Set of properties to return. Defaults to noAcl, unless the object resource specifies the acl property, when it defaults to full.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Object):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_insert_resumable(connection, bucket, upload_type, opts \\ [])

storage_objects_insert_resumable(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, nil} | {:error, Tesla.Env.t()}

Stores a new object and metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

upload_type (String.t): Upload type. Must be "resumable".

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:contentEncoding (String.t): If set, sets the contentEncoding property of the final object to this value. Setting this parameter is equivalent to setting the contentEncoding metadata property. This can be useful when uploading an object with uploadType=media to indicate the encoding of the content being uploaded.

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:kmsKeyName (String.t): Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object. Overrides the object metadata's kms_key_name value, if any. Limited availability; usable only by enabled projects.

:name (String.t): Name of the object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

:predefinedAcl (String.t): Apply a predefined set of access controls to this object.

:projection (String.t): Set of properties to return. Defaults to noAcl, unless the object resource specifies the acl property, when it defaults to full.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Object):

Returns

{:ok, %{}} on success {:error, info} on failure

storage_objects_insert_simple(connection, bucket, upload_type, metadata, data, opts \\ [])

storage_objects_insert_simple(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  GoogleApi.Storage.V1.Model.Object.t(),
  String.t(),
  keyword()
) :: {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Stores a new object and metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

upload_type (String.t): Upload type. Must be "multipart".

metadata (Object): Object metadata.

data (String.t): The file to upload.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:contentEncoding (String.t): If set, sets the contentEncoding property of the final object to this value. Setting this parameter is equivalent to setting the contentEncoding metadata property. This can be useful when uploading an object with uploadType=media to indicate the encoding of the content being uploaded.

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:kmsKeyName (String.t): Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object. Overrides the object metadata's kms_key_name value, if any. Limited availability; usable only by enabled projects.

:name (String.t): Name of the object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

:predefinedAcl (String.t): Apply a predefined set of access controls to this object.

:projection (String.t): Set of properties to return. Defaults to noAcl, unless the object resource specifies the acl property, when it defaults to full.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_list(connection, bucket, opts \\ [])

storage_objects_list(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Objects.t()} | {:error, Tesla.Env.t()}

Retrieves a list of objects matching the criteria.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which to look for objects.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:delimiter (String.t): Returns results in a directory-like mode. items will contain only objects whose names, aside from the prefix, do not contain delimiter. Objects whose names, aside from the prefix, contain delimiter will have their name, truncated after the delimiter, returned in prefixes. Duplicate prefixes are omitted.

:includeTrailingDelimiter (boolean()): If true, objects that end in exactly one instance of delimiter will have their metadata included in items in addition to prefixes.

:maxResults (integer()): Maximum number of items plus prefixes to return in a single page of responses. As duplicate prefixes are omitted, fewer total results may be returned than requested. The service will use this parameter or 1,000 items, whichever is smaller.

:pageToken (String.t): A previously-returned page token representing part of the larger set of results to view.

:prefix (String.t): Filter results to objects whose names begin with this prefix.

:projection (String.t): Set of properties to return. Defaults to noAcl.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:versions (boolean()): If true, lists all versions of an object as distinct results. The default is false. For more information, see Object Versioning.

Returns

{:ok, %GoogleApi.Storage.V1.Model.Objects{}} on success {:error, info} on failure

storage_objects_patch(connection, bucket, object, opts \\ [])

storage_objects_patch(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Patches an object's metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, selects a specific revision of this object (as opposed to the latest version, the default).

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:predefinedAcl (String.t): Apply a predefined set of access controls to this object.

:projection (String.t): Set of properties to return. Defaults to full.

:userProject (String.t): The project to be billed for this request, for Requester Pays buckets.

:body (Object):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_rewrite(connection, source_bucket, source_object, destination_bucket, destination_object, opts \\ [])

storage_objects_rewrite(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.RewriteResponse.t()}
  | {:error, Tesla.Env.t()}

Rewrites a source object to a destination object. Optionally overrides metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

source_bucket (String.t): Name of the bucket in which to find the source object.

source_object (String.t): Name of the source object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

destination_bucket (String.t): Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

destination_object (String.t): Name of the new object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:destinationKmsKeyName (String.t): Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object. Overrides the object metadata's kms_key_name value, if any.

:destinationPredefinedAcl (String.t): Apply a predefined set of access controls to the destination object.

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the destination object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the destination object's current metageneration does not match the given value.

:ifSourceGenerationMatch (String.t): Makes the operation conditional on whether the source object's current generation matches the given value.

:ifSourceGenerationNotMatch (String.t): Makes the operation conditional on whether the source object's current generation does not match the given value.

:ifSourceMetagenerationMatch (String.t): Makes the operation conditional on whether the source object's current metageneration matches the given value.

:ifSourceMetagenerationNotMatch (String.t): Makes the operation conditional on whether the source object's current metageneration does not match the given value.

:maxBytesRewrittenPerCall (String.t): The maximum number of bytes that will be rewritten per rewrite request. Most callers shouldn't need to specify this parameter - it is primarily in place to support testing. If specified the value must be an integral multiple of 1 MiB (1048576). Also, this only applies to requests where the source and destination span locations and/or storage classes. Finally, this value must not change across rewrite calls else you'll get an error that the rewriteToken is invalid.

:projection (String.t): Set of properties to return. Defaults to noAcl, unless the object resource specifies the acl property, when it defaults to full.

:rewriteToken (String.t): Include this field (from the previous rewrite response) on each rewrite request after the first one, until the rewrite response 'done' flag is true. Calls that provide a rewriteToken can omit all other request fields, but if included those fields must match the values provided in the first rewrite request.

:sourceGeneration (String.t): If present, selects a specific revision of the source object (as opposed to the latest version, the default).

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Object):

Returns

{:ok, %GoogleApi.Storage.V1.Model.RewriteResponse{}} on success {:error, info} on failure

storage_objects_set_iam_policy(connection, bucket, object, opts \\ [])

storage_objects_set_iam_policy(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, GoogleApi.Storage.V1.Model.Policy.t()} | {:error, Tesla.Env.t()}

Updates an IAM policy for the specified object.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, selects a specific revision of this object (as opposed to the latest version, the default).

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Policy):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Policy{}} on success {:error, info} on failure

storage_objects_test_iam_permissions(connection, bucket, object, permissions, opts \\ [])

storage_objects_test_iam_permissions(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  [String.t()],
  keyword()
) ::
  {:ok, GoogleApi.Storage.V1.Model.TestIamPermissionsResponse.t()}
  | {:error, Tesla.Env.t()}

Tests a set of permissions on the given object to see which, if any, are held by the caller.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

permissions ([String.t]): Permissions to test.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, selects a specific revision of this object (as opposed to the latest version, the default).

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

Returns

{:ok, %GoogleApi.Storage.V1.Model.TestIamPermissionsResponse{}} on success {:error, info} on failure

storage_objects_update(connection, bucket, object, opts \\ [])

storage_objects_update(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Object.t()} | {:error, Tesla.Env.t()}

Updates an object's metadata.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which the object resides.

object (String.t): Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:generation (String.t): If present, selects a specific revision of this object (as opposed to the latest version, the default).

:ifGenerationMatch (String.t): Makes the operation conditional on whether the object's current generation matches the given value. Setting to 0 makes the operation succeed only if there are no live versions of the object.

:ifGenerationNotMatch (String.t): Makes the operation conditional on whether the object's current generation does not match the given value. If no live object exists, the precondition fails. Setting to 0 makes the operation succeed only if there is a live version of the object.

:ifMetagenerationMatch (String.t): Makes the operation conditional on whether the object's current metageneration matches the given value.

:ifMetagenerationNotMatch (String.t): Makes the operation conditional on whether the object's current metageneration does not match the given value.

:predefinedAcl (String.t): Apply a predefined set of access controls to this object.

:projection (String.t): Set of properties to return. Defaults to full.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:body (Object):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Object{}} on success {:error, info} on failure

storage_objects_watch_all(connection, bucket, opts \\ [])

storage_objects_watch_all(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, GoogleApi.Storage.V1.Model.Channel.t()} | {:error, Tesla.Env.t()}

Watch for changes on all objects in a bucket.

Parameters

connection (GoogleApi.Storage.V1.Connection): Connection to server

bucket (String.t): Name of the bucket in which to look for objects.

opts (KeywordList): [optional] Optional parameters

:alt (String.t): Data format for the response.

:fields (String.t): Selector specifying which fields to include in a partial response.

:key (String.t): API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

:oauth_token (String.t): OAuth 2.0 token for the current user.

:prettyPrint (boolean()): Returns response with indentations and line breaks.

:quotaUser (String.t): An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

:userIp (String.t): Deprecated. Please use quotaUser instead.

:delimiter (String.t): Returns results in a directory-like mode. items will contain only objects whose names, aside from the prefix, do not contain delimiter. Objects whose names, aside from the prefix, contain delimiter will have their name, truncated after the delimiter, returned in prefixes. Duplicate prefixes are omitted.

:includeTrailingDelimiter (boolean()): If true, objects that end in exactly one instance of delimiter will have their metadata included in items in addition to prefixes.

:maxResults (integer()): Maximum number of items plus prefixes to return in a single page of responses. As duplicate prefixes are omitted, fewer total results may be returned than requested. The service will use this parameter or 1,000 items, whichever is smaller.

:pageToken (String.t): A previously-returned page token representing part of the larger set of results to view.

:prefix (String.t): Filter results to objects whose names begin with this prefix.

:projection (String.t): Set of properties to return. Defaults to noAcl.

:userProject (String.t): The project to be billed for this request. Required for Requester Pays buckets.

:versions (boolean()): If true, lists all versions of an object as distinct results. The default is false. For more information, see Object Versioning.

:resource (Channel):

Returns

{:ok, %GoogleApi.Storage.V1.Model.Channel{}} on success {:error, info} on failure