google_api_storage v0.1.0 GoogleApi.Storage.V1.Model.BucketAccessControl

An access-control entry.

Attributes

bucket (String.t): The name of the bucket. Defaults to: null.

domain (String.t): The domain associated with the entity, if any. Defaults to: null.

email (String.t): The email address associated with the entity, if any. Defaults to: null.

entity (String.t): The entity holding the permission, in one of the following forms: - user-userId - user-email - group-groupId - group-email - domain-domain - project-team-projectId - allUsers - allAuthenticatedUsers Examples: - The user liz@example.com would be user-liz@example.com. - The group example@googlegroups.com would be group-example@googlegroups.com. - To refer to all members of the Google Apps for Business domain example.com, the entity would be domain-example.com. Defaults to: null.

entityId (String.t): The ID for the entity, if any. Defaults to: null.

etag (String.t): HTTP 1.1 Entity tag for the access-control entry. Defaults to: null.

id (String.t): The ID of the access-control entry. Defaults to: null.

kind (String.t): The kind of item this is. For bucket access control entries, this is always storage#bucketAccessControl. Defaults to: null.

projectTeam (BucketAccessControlProjectTeam): Defaults to: null.

role (String.t): The access permission for the entity. Defaults to: null.

selfLink (String.t): The link to this access-control entry. Defaults to: null.

Types

t()

t() :: %GoogleApi.Storage.V1.Model.BucketAccessControl{
  bucket: any(),
  domain: any(),
  email: any(),
  entity: any(),
  entityId: any(),
  etag: any(),
  id: any(),
  kind: any(),
  projectTeam: GoogleApi.Storage.V1.Model.BucketAccessControlProjectTeam.t(),
  role: any(),
  selfLink: any()
}

Functions

decode(value, options)

decode(struct(), keyword()) :: struct()

Unwrap a decoded JSON object into its complex fields.