google_api_storage v0.1.0 GoogleApi.Storage.V1.Model.BucketLifecycleCondition

The condition(s) under which the action will be taken.

Attributes

age (integer()): Age of an object (in days). This condition is satisfied when an object reaches the specified age. Defaults to: null.

createdBefore (Date.t): A date in RFC 3339 format with only the date part (for instance, "2013-01-15"). This condition is satisfied when an object is created before midnight of the specified date in UTC. Defaults to: null.

isLive (boolean()): Relevant only for versioned objects. If the value is true, this condition matches live objects; if the value is false, it matches archived objects. Defaults to: null.

matchesStorageClass ([String.t]): Objects having any of the storage classes specified by this condition will be matched. Values include MULTI_REGIONAL, REGIONAL, NEARLINE, COLDLINE, STANDARD, and DURABLE_REDUCED_AVAILABILITY. Defaults to: null.

numNewerVersions (integer()): Relevant only for versioned objects. If the value is N, this condition is satisfied when there are at least N versions (including the live version) newer than this version of the object. Defaults to: null.

Types

t()

t() :: %GoogleApi.Storage.V1.Model.BucketLifecycleCondition{
  age: any(),
  createdBefore: Date.t(),
  isLive: any(),
  matchesStorageClass: [any()],
  numNewerVersions: any()
}

Functions

decode(value, options)

decode(struct(), keyword()) :: struct()

Unwrap a decoded JSON object into its complex fields.