google_api_storage v0.1.0 GoogleApi.Storage.V1.Model.Channel

An notification channel used to watch for resource changes.

Attributes

address (String.t): The address where notifications are delivered for this channel. Defaults to: null.

expiration (String.t): Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds. Optional. Defaults to: null.

id (String.t): A UUID or similar unique string that identifies this channel. Defaults to: null.

kind (String.t): Identifies this as a notification channel used to watch for changes to a resource. Value: the fixed string "api#channel". Defaults to: null.

params (%{optional(String.t) => String.t}): Additional parameters controlling delivery channel behavior. Optional. Defaults to: null.

payload (boolean()): A Boolean value to indicate whether payload is wanted. Optional. Defaults to: null.

resourceId (String.t): An opaque ID that identifies the resource being watched on this channel. Stable across different API versions. Defaults to: null.

resourceUri (String.t): A version-specific identifier for the watched resource. Defaults to: null.

token (String.t): An arbitrary string delivered to the target address with each notification delivered over this channel. Optional. Defaults to: null.

type (String.t): The type of delivery mechanism used for this channel. Defaults to: null.

Types

t()

t() :: %GoogleApi.Storage.V1.Model.Channel{
  address: any(),
  expiration: any(),
  id: any(),
  kind: any(),
  params: map(),
  payload: any(),
  resourceId: any(),
  resourceUri: any(),
  token: any(),
  type: any()
}

Functions

decode(value, options)

decode(struct(), keyword()) :: struct()

Unwrap a decoded JSON object into its complex fields.