google_api_storage v0.1.0 GoogleApi.Storage.V1.Model.Notification

A subscription to receive Google PubSub notifications.

Attributes

custom_attributes (%{optional(String.t) => String.t}): An optional list of additional attributes to attach to each Cloud PubSub message published for this notification subscription. Defaults to: null.

etag (String.t): HTTP 1.1 Entity tag for this subscription notification. Defaults to: null.

event_types ([String.t]): If present, only send notifications about listed event types. If empty, sent notifications for all event types. Defaults to: null.

id (String.t): The ID of the notification. Defaults to: null.

kind (String.t): The kind of item this is. For notifications, this is always storage#notification. Defaults to: null.

object_name_prefix (String.t): If present, only apply this notification configuration to object names that begin with this prefix. Defaults to: null.

payload_format (String.t): The desired content of the Payload. Defaults to: null.

selfLink (String.t): The canonical URL of this notification. Defaults to: null.

topic (String.t): The Cloud PubSub topic to which this subscription publishes. Formatted as: '//pubsub.googleapis.com/projects/{project-identifier}/topics/{my-topic}' Defaults to: null.

Types

t()

t() :: %GoogleApi.Storage.V1.Model.Notification{
  custom_attributes: map(),
  etag: any(),
  event_types: [any()],
  id: any(),
  kind: any(),
  object_name_prefix: any(),
  payload_format: any(),
  selfLink: any(),
  topic: any()
}

Functions

decode(value, options)

decode(struct(), keyword()) :: struct()

Unwrap a decoded JSON object into its complex fields.