google_api_storage v0.1.0 GoogleApi.Storage.V1.Model.Object

An object.

Attributes

acl ([ObjectAccessControl]): Access controls on the object. Defaults to: null.

bucket (String.t): The name of the bucket containing this object. Defaults to: null.

cacheControl (String.t): Cache-Control directive for the object data. If omitted, and the object is accessible to all anonymous users, the default will be public, max-age=3600. Defaults to: null.

componentCount (integer()): Number of underlying components that make up this object. Components are accumulated by compose operations. Defaults to: null.

contentDisposition (String.t): Content-Disposition of the object data. Defaults to: null.

contentEncoding (String.t): Content-Encoding of the object data. Defaults to: null.

contentLanguage (String.t): Content-Language of the object data. Defaults to: null.

contentType (String.t): Content-Type of the object data. If an object is stored without a Content-Type, it is served as application/octet-stream. Defaults to: null.

crc32c (String.t): CRC32c checksum, as described in RFC 4960, Appendix B; encoded using base64 in big-endian byte order. For more information about using the CRC32c checksum, see Hashes and ETags: Best Practices. Defaults to: null.

customerEncryption (ObjectCustomerEncryption): Defaults to: null.

etag (String.t): HTTP 1.1 Entity tag for the object. Defaults to: null.

eventBasedHold (boolean()): Whether an object is under event-based hold. Event-based hold is a way to retain objects until an event occurs, which is signified by the hold's release (i.e. this value is set to false). After being released (set to false), such objects will be subject to bucket-level retention (if any). One sample use case of this flag is for banks to hold loan documents for at least 3 years after loan is paid in full. Here, bucket-level retention is 3 years and the event is the loan being paid in full. In this example, these objects will be held intact for any number of years until the event has occurred (event-based hold on the object is released) and then 3 more years after that. That means retention duration of the objects begins from the moment event-based hold transitioned from true to false. Defaults to: null.

generation (String.t): The content generation of this object. Used for object versioning. Defaults to: null.

id (String.t): The ID of the object, including the bucket name, object name, and generation number. Defaults to: null.

kind (String.t): The kind of item this is. For objects, this is always storage#object. Defaults to: null.

kmsKeyName (String.t): Cloud KMS Key used to encrypt this object, if the object is encrypted by such a key. Limited availability; usable only by enabled projects. Defaults to: null.

md5Hash (String.t): MD5 hash of the data; encoded using base64. For more information about using the MD5 hash, see Hashes and ETags: Best Practices. Defaults to: null.

mediaLink (String.t): Media download link. Defaults to: null.

metadata (%{optional(String.t) => String.t}): User-provided metadata, in key/value pairs. Defaults to: null.

metageneration (String.t): The version of the metadata for this object at this generation. Used for preconditions and for detecting changes in metadata. A metageneration number is only meaningful in the context of a particular generation of a particular object. Defaults to: null.

name (String.t): The name of the object. Required if not specified by URL parameter. Defaults to: null.

owner (ObjectOwner): Defaults to: null.

retentionExpirationTime (DateTime.t): A server-determined value that specifies the earliest time that the object's retention period expires. This value is in RFC 3339 format. Note 1: This field is not provided for objects with an active event-based hold, since retention expiration is unknown until the hold is removed. Note 2: This value can be provided even when temporary hold is set (so that the user can reason about policy without having to first unset the temporary hold). Defaults to: null.

selfLink (String.t): The link to this object. Defaults to: null.

size (String.t): Content-Length of the data in bytes. Defaults to: null.

storageClass (String.t): Storage class of the object. Defaults to: null.

temporaryHold (boolean()): Whether an object is under temporary hold. While this flag is set to true, the object is protected against deletion and overwrites. A common use case of this flag is regulatory investigations where objects need to be retained while the investigation is ongoing. Note that unlike event-based hold, temporary hold does not impact retention expiration time of an object. Defaults to: null.

timeCreated (DateTime.t): The creation time of the object in RFC 3339 format. Defaults to: null.

timeDeleted (DateTime.t): The deletion time of the object in RFC 3339 format. Will be returned if and only if this version of the object has been deleted. Defaults to: null.

timeStorageClassUpdated (DateTime.t): The time at which the object's storage class was last changed. When the object is initially created, it will be set to timeCreated. Defaults to: null.

updated (DateTime.t): The modification time of the object metadata in RFC 3339 format. Defaults to: null.

Types

t()

t() :: %GoogleApi.Storage.V1.Model.Object{
  acl: [GoogleApi.Storage.V1.Model.ObjectAccessControl.t()],
  bucket: any(),
  cacheControl: any(),
  componentCount: any(),
  contentDisposition: any(),
  contentEncoding: any(),
  contentLanguage: any(),
  contentType: any(),
  crc32c: any(),
  customerEncryption: GoogleApi.Storage.V1.Model.ObjectCustomerEncryption.t(),
  etag: any(),
  eventBasedHold: any(),
  generation: any(),
  id: any(),
  kind: any(),
  kmsKeyName: any(),
  md5Hash: any(),
  mediaLink: any(),
  metadata: map(),
  metageneration: any(),
  name: any(),
  owner: GoogleApi.Storage.V1.Model.ObjectOwner.t(),
  retentionExpirationTime: DateTime.t(),
  selfLink: any(),
  size: any(),
  storageClass: any(),
  temporaryHold: any(),
  timeCreated: DateTime.t(),
  timeDeleted: DateTime.t(),
  timeStorageClassUpdated: DateTime.t(),
  updated: DateTime.t()
}

Functions

decode(value, options)

decode(struct(), keyword()) :: struct()

Unwrap a decoded JSON object into its complex fields.