google_api_storage v0.1.0 GoogleApi.Storage.V1.Model.Policy

A bucket/object IAM policy.

Attributes

bindings ([PolicyBindings]): An association between a role, which comes with a set of permissions, and members who may assume that role. Defaults to: null.

etag (binary()): HTTP 1.1 Entity tag for the policy. Defaults to: null.

kind (String.t): The kind of item this is. For policies, this is always storage#policy. This field is ignored on input. Defaults to: null.

resourceId (String.t): The ID of the resource to which this policy belongs. Will be of the form projects//buckets/bucket for buckets, and projects//buckets/bucket/objects/object for objects. A specific generation may be specified by appending #generationNumber to the end of the object name, e.g. projects/_/buckets/my-bucket/objects/data.txt#17. The current generation can be denoted with #0. This field is ignored on input. Defaults to: null.

Types

t()

t() :: %GoogleApi.Storage.V1.Model.Policy{
  bindings: [GoogleApi.Storage.V1.Model.PolicyBindings.t()],
  etag: any(),
  kind: any(),
  resourceId: any()
}

Functions

decode(value, options)

decode(struct(), keyword()) :: struct()

Unwrap a decoded JSON object into its complex fields.